server/requirements.txt
newspaper3k
google-generativeai
python-dotenv
fastapi
uvicorn
lxml_html_clean
youtube-transcript-api